+++ CDJSVis/filtering/filterer.py
@@ -252,6 +252,9 @@ class Filterer(object):
             elif filterName == "Bond order":
                 self.bondOrderFilter(filterSettings)
             
+            elif filterName == "Atom index":
+                self.atomIndexFilter(filterSettings)
+            
             # write to log
             if self.parent.defectFilterSelected:
                 self.NVis = len(interstitials) + len(vacancies) + len(antisites) + len(splitInterstitials)
@@ -806,6 +809,26 @@ class Filterer(object):
             # store scalars
             scalars.resize(NVisible, refcheck=False)
             self.scalarsDict["Displacement"] = scalars
+    
+    def atomIndexFilter(self, settings):
+        """
+        Atom index filter
+        
+        """
+        lattice = self.pipelinePage.inputState
+        
+        # old scalars arrays (resize as appropriate)
+        NScalars, fullScalars = self.makeFullScalarsArray()
+        
+        # run displacement filter
+        NVisible = filtering_c.atomIndexFilter(self.visibleAtoms, lattice.atomID, settings.filteringEnabled, settings.minVal, settings.maxVal, 
+                                               NScalars, fullScalars)
+        
+        # update scalars dict
+        self.storeFullScalarsArray(NVisible, NScalars, fullScalars)
+        
+        # resize visible atoms
+        self.visibleAtoms.resize(NVisible, refcheck=False)
     
     def cropFilter(self, settings):
         """